refactor(app): extract ManageBackups inventory + job composables (#399)
Reduce ManageBackups.vue from 1115 to 579 lines by extracting cohesive
seams into view-adjacent modules, keeping behavior, routes, props, emits,
and network calls identical:

- useBackupInventory: list state, filters, sort, pagination, fetch +
  download, plus pure formatters (file size/date/type/badge).
- useBackupJobs: backup/restore async-job trackers (reusing useAsyncJob),
  confirm-modal state, and trigger/confirm/delete actions with toasts.
- BackupConfirmModal: shared restore/delete confirmation modal (DRY the
  two near-identical danger panels); uses v-model so @hidden resets the
  typed-confirm input on every close path.

Adds a unit test for the new pure-helper + filter/pagination seam.

Advances #399. Part of #346.
